feat(api): apply errorWrapper on each api handler
With error wrapper, each handle can throw an ApiError object if the casus of error
is known, then the handler will log the error and send response with data in ApiError;
error not created by app (not ApiError type) will be catched and send a response with
status 500.

New APIs in future can add this hanlder, too.