feat(api): apply session and its validator on each api handler
Modify the createSessionWrapper to withSession, which accept a next handler and returns a function receive
req and res as args. So it can use other middleware as a argument.

Add a new function createSessionValidator which can create a middleware for validate session. The validator
can be used on handler of any api endpoint.